rename to public_goods_games and point tests at it
