+++ .github/workflows/ci.yml
@@ -21,3 +21,4 @@ jobs:
       - name: Run ruff
         run: |
           uv run ruff check
+g
