+++ src/public_goods_games/contribution_rules.py
